feat(cli): add generate-data command
Move offline hidden-states generation from
scripts/data_generation_offline.py into speculators.cli.generate_data,
wire as `speculators generate-data` under the Pipeline help panel.
Original script becomes a deprecation shim.

Part of vllm-project/speculators#927 (Sub-PR 4 of 6)
